requirements.txt
boto3==1.20.17
botocore==1.23.54
confluent-kafka==2.2.0
falcon==3.0.1
falcon-auth==1.1.0
falcon-cors==1.1.7
falcon-multipart==0.2.0
jmespath==0.10.0
middleware==1.2.3
numpy==1.21.4
protobuf==4.21.4
psycopg-binary==3.1.8
psycopg2-binary==2.9.3
py4j==0.10.9
PyJWT==2.4.0
pyspark==3.1.3
python-dateutil==2.8.2
python-dotenv==0.19.2
python-json-logger==2.0.2
pytz==2022.1
s3transfer==0.5.2
six==1.16.0
urllib3==1.26.11
waitress==2.0.0
mysql-connector-python==8.0.27
requests==2.28.1
futures3==1.0.0
pandas==1.5.3
elastic-apm==6.15.1
SQLAlchemy==2.0.10
connectorx==0.3.1
pyarrow==12.0.0
